build(backend): added starter setup for the backend
TODO: fix the issue with noEmit this way we would be able to build the server without the needs to change the config file
