+++ src/apps/server/README.md
@@ -9,4 +9,10 @@ pnpm install
 tsx index.ts
 ```
 
-to test the api routes open [http://localhost:4000/panel](http://localhost:4000/panel)
+To test the api routes open [http://localhost:4000/panel](http://localhost:4000/panel)
+
+notes:
+
+To build the application in production we should first compile Typescript code
+and the run the .js using node 
+[see details here https://github.com/TypeStrong/ts-node/issues/104](https://github.com/TypeStrong/ts-node/issues/104)
