+++ src/apps/server/nodemon.json
@@ -1,0 +1,6 @@
+{
+  "watch": ["./*"],
+  "ext": "ts",
+  "ignore": ["./*.spec.ts"],
+  "exec": "ts-node ./index.ts"
+}
