+++ src/apps/server/package.json
@@ -4,8 +4,8 @@
   "private": true,
   "type": "module",
   "scripts": {
-    "dev": "tsx index.ts",
-    "build": "echo 'Add build script here'",
+    "dev": "nodemon",
+    "build": "tsc  --project tsconfig.json",
     "test": "echo 'Add test script here'",
     "lint": "echo 'Add lint script here'"
   },
@@ -18,8 +18,9 @@
     "@trpc/server": "^10.45.2",
     "@types/cors": "^2.8.18",
     "@types/express": "^5.0.2",
+    "nodemon": "^3.1.10",
     "trpc-panel": "^1.3.4",
-    "tsx": "^4.19.4",
+    "ts-node": "^10.9.2",
     "typescript": "5.5.4"
   }
 }
