+++ src/package.json
@@ -4,13 +4,19 @@
     "build": "turbo run build",
     "dev": "turbo run dev",
     "lint": "turbo run lint",
-    "format": "prettier --write \"**/*.{ts,tsx,md}\""
+    "format": "prettier --write \"**/*.{ts,tsx,md}\"",
+    "graph": "turbo-graph"
   },
   "devDependencies": {
+    "@yeger/turbo-graph": "^2.0.13",
     "eslint": "^8.57.0",
     "prettier": "^3.2.5",
+    "tailwindcss": "^4.1.8",
     "turbo": "^2.5.3"
   },
   "packageManager": "pnpm@8.15.6",
-  "name": "with-vite-react"
+  "name": "with-vite-react",
+  "dependencies": {
+    "pnpm": "^10.11.0"
+  }
 }
